+++ CHANGELOG.md
@@ -23,6 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.2] - 2026-03-07
 
+### Added
+
+- Add refresh button and handler to re-fetch inventory items in MainContainer
+
 ### Changed
 
 - Refactor MainContainer to prevent unnecessary Airtable refetches by tracking last-fetched parameters